refactor: ai chat controller로 분리
