+++ app/api/routes/health_check_controller.py
@@ -1,9 +1,7 @@
 from typing import Any
 
-from fastapi import APIRouter, HTTPException
-from pydantic import BaseModel
+from fastapi import APIRouter
 
-from app.api.service.rag_pipeline import rag_pipeline
 
 router = APIRouter()
 
@@ -11,23 +9,3 @@ router = APIRouter()
 @router.get("")
 def health_check() -> Any:
     return "success"
-
-
-class QueryRequest(BaseModel):
-    query: str
-
-
-class QueryResponse(BaseModel):
-    rag_result: str
-    sources: list
-
-
-@router.post("/query", response_model=QueryResponse)
-async def get_query_result(request: QueryRequest):
-    try:
-        result = rag_pipeline.query(request.query)
-        return result
-    except Exception as e:
-        # 에러 핸들링
-        raise HTTPException(status_code=500, detail=str(e))
-
